+++ src/pages/Gallery/components/MenuBar/MenuBar.styles.ts
