+++ src/pages/Gallery/components/MenuBar/MenuBar.types.ts
@@ -1,0 +1,6 @@
+import type { Dispatch, SetStateAction } from 'react'
+import type { MenuProps } from '../../gallery.types'
+
+export interface MenuBarProps {
+  setSelectedCategory: Dispatch<SetStateAction<MenuProps>>
+}
