+++ src/pages/Gallery/components/MenuBar/index.ts
@@ -1,0 +1,1 @@
+export * from './MenuBar'
