+++ src/pages/Gallery/components/PictureItem/index.ts
@@ -1,0 +1,1 @@
+export * from './PictureItem'
