+++ src/pages/Gallery/components/PictureList/index.ts
@@ -1,0 +1,1 @@
+export * from './PictureList'
