+++ src/pages/Gallery/components/index.ts
@@ -1,0 +1,3 @@
+export * from './MenuBar'
+export * from './PictureItem'
+export * from './PictureList'
